+++ pkg/service/did/model.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/TBD54566975/ssi-sdk/crypto"
 	didsdk "github.com/TBD54566975/ssi-sdk/did"
+	"github.com/TBD54566975/ssi-sdk/did/ion"
 	"github.com/TBD54566975/ssi-sdk/did/resolution"
 	"github.com/tbd54566975/ssi-service/pkg/service/common"
 )
@@ -84,3 +85,26 @@ type DeleteDIDRequest struct {
 	Method didsdk.Method `json:"method" validate:"required"`
 	ID     string        `json:"id" validate:"required"`
 }
+
+type UpdateIONDIDRequest struct {
+	DID ion.ION `json:"did"`
+
+	StateChange ion.StateChange `json:"stateChange"`
+}
+
+type UpdateIONDIDResponse struct {
+	DID didsdk.Document `json:"did"`
+}
+
+type UpdateRequestStatus string
+
+func (s UpdateRequestStatus) Bytes() []byte {
+	return []byte(s)
+}
+
+const (
+	PreAnchorStatus   UpdateRequestStatus = "pre-anchor"
+	AnchorErrorStatus UpdateRequestStatus = "anchor-error"
+	AnchoredStatus    UpdateRequestStatus = "anchored"
+	DoneStatus        UpdateRequestStatus = "done"
+)
